feat: Refactor Settings page to use tabs for navigation and integrate new panels
- Implemented tab navigation for the Settings page, allowing users to switch between Profile, Notifications, Feature Flags, Translations, Languages, File Manager, Security, and Maintenance sections.
- Created separate panels for Feature Flags, Translations, Languages, Notifications, and Maintenance to encapsulate functionality and improve code organization.
- Updated the Translations page to utilize the new TranslationsPanel component.
- Added PanelHeader component to standardize headers across different panels.
- Enhanced user experience by managing state and search functionality within each panel.
