+++ frontend/src/views/Settings/pages/Translations.tsx
@@ -1,60 +1,10 @@
-import { useState } from "react";
 import { PageLayout } from "@/components/PageLayout";
-import { PageHeader } from "@/components/PageHeader";
-import { TranslationsCard } from "../components/TranslationsCard";
-import { TranslationDialog } from "../components/TranslationDialog";
-import { Translation } from "@/interfaces/translation.interface";
+import { TranslationsPanel } from "../panels/TranslationsPanel";
 
 export function Translations() {
-  const [searchQuery, setSearchQuery] = useState("");
-  const [refreshKey, setRefreshKey] = useState(0);
-  const [isDialogOpen, setIsDialogOpen] = useState(false);
-  const [dialogMode, setDialogMode] = useState<"create" | "edit">("create");
-  const [translationToEdit, setTranslationToEdit] = useState<Translation | null>(null);
-
-  const handleRefresh = () => {
-    setRefreshKey((prev) => prev + 1);
-  };
-
-  const handleOpenCreate = () => {
-    setDialogMode("create");
-    setTranslationToEdit(null);
-    setIsDialogOpen(true);
-  };
-
-  const handleOpenEdit = (translation: Translation | null, mode: "create" | "edit") => {
-    setDialogMode(mode);
-    setTranslationToEdit(translation);
-    setIsDialogOpen(true);
-  };
-
   return (
     <PageLayout>
-      <PageHeader
-        title="Translations"
-        subtitle="View and manage application translations"
-        searchQuery={searchQuery}
-        onSearchChange={setSearchQuery}
-        searchPlaceholder="Search translations..."
-        actionButtonText="Add Translation"
-        onActionClick={handleOpenCreate}
-      />
-
-      <TranslationsCard
-        searchQuery={searchQuery}
-        refreshKey={refreshKey}
-        onEditTranslation={handleOpenEdit}
-        onRefresh={handleRefresh}
-        onAddTranslation={handleOpenCreate}
-      />
-
-      <TranslationDialog
-        isOpen={isDialogOpen}
-        onOpenChange={setIsDialogOpen}
-        mode={dialogMode}
-        translationToEdit={translationToEdit}
-        onTranslationSaved={handleRefresh}
-      />
+      <TranslationsPanel variant="page" />
     </PageLayout>
   );
 }
