+++ frontend/src/views/Settings/panels/LanguagesPanel.tsx
@@ -1,0 +1,64 @@
+import { useState } from "react";
+import { Language } from "@/interfaces/translation.interface";
+import { LanguagesCard } from "../components/LanguagesCard";
+import { LanguageDialog } from "../components/LanguageDialog";
+import { PanelHeader, type PanelVariant } from "./PanelHeader";
+
+/**
+ * Self-contained Languages experience (header + table + dialog). Rendered both
+ * by the standalone `/settings/languages` route (variant="page") and by the
+ * Languages tab on the Settings page (variant="tab").
+ */
+export function LanguagesPanel({ variant = "page" }: { variant?: PanelVariant }) {
+  const [searchQuery, setSearchQuery] = useState("");
+  const [isDialogOpen, setIsDialogOpen] = useState(false);
+  const [refreshKey, setRefreshKey] = useState(0);
+  const [dialogMode, setDialogMode] = useState<"create" | "edit">("create");
+  const [languageToEdit, setLanguageToEdit] = useState<Language | null>(null);
+
+  const handleLanguageSaved = () => {
+    setRefreshKey((prevKey) => prevKey + 1);
+  };
+
+  const handleCreateLanguage = () => {
+    setDialogMode("create");
+    setLanguageToEdit(null);
+    setIsDialogOpen(true);
+  };
+
+  const handleEditLanguage = (language: Language) => {
+    setDialogMode("edit");
+    setLanguageToEdit(language);
+    setIsDialogOpen(true);
+  };
+
+  return (
+    <div className="space-y-6">
+      <PanelHeader
+        variant={variant}
+        title="Languages"
+        subtitle="View and manage supported languages for translations"
+        searchQuery={searchQuery}
+        onSearchChange={setSearchQuery}
+        searchPlaceholder="Search languages..."
+        actionButtonText="Add New Language"
+        onActionClick={handleCreateLanguage}
+      />
+
+      <LanguagesCard
+        searchQuery={searchQuery}
+        refreshKey={refreshKey}
+        onEditLanguage={handleEditLanguage}
+        onCreateLanguage={handleCreateLanguage}
+      />
+
+      <LanguageDialog
+        isOpen={isDialogOpen}
+        onOpenChange={setIsDialogOpen}
+        onLanguageSaved={handleLanguageSaved}
+        languageToEdit={languageToEdit}
+        mode={dialogMode}
+      />
+    </div>
+  );
+}
